Split education credit variables into separate files
